Several updates
Changed ground detection logic to register when the player is not on the
ground, and disable jumping if so. Uncomented the acceleration and speed
call for being on the ground. Not sure the effects of taht, I think I
uncommented them for a jump bug that i was no longer seeing.

Cleaned up the player object to place getters and setters at the bottom
and added some comments.

Changed speed of proctiles.